backend/whisper_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai-whisper==20231117
yt-dlp==2024.3.10
ffmpeg-python==0.2.0
pydantic==2.6.1
python-multipart==0.0.9 